ReaderController refactoring of Post method which is responsible of deleting newsletter and appropriate View.

+++ RSS_Reader/Controllers/ReaderController.cs
@@ -11,6 +11,7 @@ namespace RSS_Reader.Controllers
     public class ReaderController : Controller
     {
         private IRssDocumentsRepository _context;
+        private string CombGlob = "";
 
         // GET: Reader
         public ActionResult Index()
@@ -25,8 +26,12 @@ namespace RSS_Reader.Controllers
         }
 
         [HttpPost]
-        public ActionResult CancelNewsletter(string EmailAddress)
+        public ActionResult CancelNewsletter(string EmailAddress, string Combination)
         {
+            if (Combination != CombGlob)
+            {
+                return RedirectToAction("ConfirmationOfCancellingNewsletter", new { emailAddress = EmailAddress });
+            }
             _context = new RssDocumentsRepository();
             _context.DeleteFromMailingList(EmailAddress);
 
@@ -47,6 +52,7 @@ namespace RSS_Reader.Controllers
             }
 
             ViewBag.combinationString = identityChecker;
+            CombGlob = identityChecker;
 
             return View();
         }
